Getting item data through useShowItems.js hook. Displaying dropped item on hover.
